% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dailyexposure.R
\name{dailyexposure}
\alias{dailyexposure}
\title{Daily noise exposure level}
\usage{
dailyexposure(level, expositiontime, referencetime = 8)
}
\arguments{
\item{level}{Noise level.}

\item{expositiontime}{Time of noise exposition.}

\item{referencetime}{Time of reference (default = 8).}
}
\description{
Calculates the standardized 8-hour equivalent daily level.
}
\details{
Calculates the standardized 8-hour equivalent daily level.
According to Directive 2003/10/EC and ISO 1999 standard.
}
