% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noisedosis.R
\name{noisedosis}
\alias{noisedosis}
\title{Noise dosis}
\usage{
noisedosis(level, expositiontime, referencelevel = 85, exchangerate = 3)
}
\arguments{
\item{level}{Vector of dB(A) levels.}

\item{expositiontime}{Time of noise exposition.}

\item{referencelevel}{Level of reference (default = 85).}

\item{exchangerate}{Exchange rate (3 dB (EU) or 5 dB (OSHA) (default = 3).}
}
\description{
Calculates the noise dosis in a especific time of exposition.
}
\details{
Calculates the standardized 8-hour equivalent daily level.
According to Directive 2003/10/EC and ISO 1999 standard.
}
