% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noiselevelshiny.R
\name{noiselevelshiny}
\alias{noiselevelshiny}
\title{Shiny for noise level calculation.}
\usage{
noiselevelshiny()
}
\description{
Calculates the noise level.
}
\details{
For launch shiny:
Write 'noiselevelshiny()'
}
