% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noisemapshiny.R
\name{noisemapshiny}
\alias{noisemapshiny}
\title{Shiny for noise risk exposure map.}
\usage{
noisemapshiny()
}
\description{
Calculates the risk exposure map.
You can manipulate several factors.
}
\details{
For launch shiny:
Write 'noisemapshiny()'
}
