% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orpnoise.R
\docType{_PACKAGE}
\name{orpnoise}
\alias{orpnoise}
\title{orpnoise}
\description{
Noise exposure in occupational risk prevention
}
\details{
Set of functions and shinys for the calculation
of noise exposure risks in work environments.
Load 'noisemapshiny()' to open the shiny on the map of auditory exposure
according to the sources.
}
\author{
Raúl Aguilar-Elena \email{raul.aguilar.e@professor.universidadviu.com}

Alejandro Guillén-Riquelme \email{aguillenr@universidadviu.com}
}
