% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soundweighting.R
\name{soundweighting}
\alias{soundweighting}
\title{A weighting frequencies calculation.}
\usage{
soundweighting(frequencies, levels, ponderation = "A")
}
\arguments{
\item{frequencies}{Vector of frequencies in Hz.}

\item{levels}{Levels without weighting (dB).}

\item{ponderacion}{Carácter que indica el tipo de ponderación a aplicar. Los valores permitidos son:
\describe{
  \item{"A"}{Ponderación A: simula la respuesta del oído humano a bajos niveles de sonido (40 fon).
             Atenúa frecuencias bajas y altas.}
  \item{"C"}{Ponderación C: simula la respuesta del oído humano a altos niveles de sonido (100 fon).
             Atenúa menos las bajas frecuencias que la ponderación A.}
  \item{"Z"}{Ponderación Cero (Zero): no aplica corrección. Equivalente a la ponderación lineal.}
}
Por defecto es "A".}
}
\description{
Calculates A weighting, using frequencial weighting as IEC 61672.
}
\details{
Calculates A weighting, using frequencial weighting as IEC 61672 standard
for acoustic measurements.
}
